add new github actions build (#10)
