+++ .gitignore
@@ -11,3 +11,8 @@ sdk/nodejs/node_modules
 *.iml
 
 yarn.lock
+
+**/schema.go
+**/version.txt
+
+ci-scripts
