+++ .golangci.yml
@@ -1,3 +1,8 @@
+run:
+  timeout: 5m
+  skip-files: 
+    - schema.go
+    - pulumiManifest.go
 linters:
   enable-all: false
   enable:
