+++ .goreleaser.prerelease.yml
@@ -1,0 +1,40 @@
+before:
+  hooks:
+    - cd provider && go mod download
+builds:
+- env:
+  - CGO_ENABLED=0
+  - GO111MODULE=on
+  goos:
+    - darwin
+    - windows
+    - linux
+  goarch:
+    - amd64
+  dir: provider
+  main: ./cmd/pulumi-resource-kong/
+  ldflags: -X github.com/pulumi/pulumi-kong/provider/pkg/version.Version={{.Tag}}
+  binary: pulumi-resource-kong
+archives:
+- name_template: "{{ .Binary }}-{{ .Tag }}-{{ .Os }}-{{ .Arch }}"
+  format_overrides:
+    - goos: windows
+      format: zip
+  replacements:
+      amd64: x64
+      386: x86
+  id: archive
+snapshot:
+  name_template: "{{ .Tag }}-SNAPSHOT"
+changelog:
+  skip: true
+release:
+  disable: true
+  prerelease: true
+blobs:
+  - provider: s3
+    region: us-west-2
+    bucket: get.pulumi.com
+    folder: releases/plugins/
+    ids:
+    - archive
