--- examples/.gitignore
@@ -1,3 +1,0 @@
-.pulumi/
-**/bin/
-node_modules/
