+++ examples/examples_py_test.go
@@ -1,0 +1,22 @@
+// Copyright 2016-2017, Pulumi Corporation.  All rights reserved.
+// +build nodejs all
+
+package examples
+
+import (
+	"path/filepath"
+	"testing"
+
+	"github.com/pulumi/pulumi/pkg/v2/testing/integration"
+)
+
+func getPythonBaseOptions(t *testing.T) integration.ProgramTestOptions {
+	base := getBaseOptions()
+	basePython := base.With(integration.ProgramTestOptions{
+		Dependencies: []string{
+			filepath.Join("..", "sdk", "python", "bin"),
+		},
+	})
+
+	return basePython
+}
