--- provider/cmd/pulumi-resource-kong/.gitignore
@@ -1,1 +1,0 @@
-schema.go
