--- sdk/.gitignore
@@ -1,4 +1,0 @@
-/nodejs/bin/
-/nodejs/node_modules/
-/python/bin/
-
